+++ packages/tracer/src/types/ProviderService.ts
@@ -40,9 +40,57 @@ interface ProviderServiceInterface {
 
   captureHTTPsGlobal(): void;
 
+  /**
+   * Instrument `fetch` requests with AWS X-Ray
+   */
+  instrumentFetch(): void;
+
   putAnnotation(key: string, value: string | number | boolean): void;
 
   putMetadata(key: string, value: unknown, namespace?: string): void;
 }
 
-export type { ProviderServiceInterface, ContextMissingStrategy };
+/**
+ * Subsegment that contains information for a request made to a remote service
+ */
+interface HttpSubsegment extends Subsegment {
+  namespace: 'remote';
+  http: {
+    request?: {
+      url: string;
+      method: string;
+    };
+    response?: {
+      status: number;
+      content_length?: number;
+    };
+  };
+}
+
+/**
+ * Partial shape of the message sent to the `undici:request:create` diagnostics channel
+ */
+type MessageOnRequestStart = {
+  request: {
+    origin: string;
+    method: string;
+  };
+};
+
+/**
+ * Partial shape of the message sent to the `undici:request:headers` diagnostics channel
+ */
+type MessageOnResponse = MessageOnRequestStart & {
+  response: {
+    statusCode: number;
+    headers: Uint8Array[];
+  };
+};
+
+export type {
+  ProviderServiceInterface,
+  ContextMissingStrategy,
+  HttpSubsegment,
+  MessageOnRequestStart,
+  MessageOnResponse,
+};
